% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gMCPmini-package.R
\docType{package}
\name{gMCPmini-package}
\alias{gMCPmini}
\alias{gMCPmini-package}
\title{gMCPmini: Graph Based Multiple Comparison Procedures (Mini Version)}
\description{
Selected Functions From gMCP package to support gsDesign and remove JAVA dependency. The package version is the same as the gMCP package version.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://allenzhuaz.github.io/gMCPmini}
  \item \url{https://github.com/allenzhuaz/gMCPmini}
}

}
\author{
\strong{Maintainer}: Yilong Zhang \email{yilong.zhang@merck.com}

Authors:
\itemize{
  \item Yalin Zhu \email{yalin.zhu@merck.com}
  \item Xuan Deng \email{xuan.deng@merck.com}
  \item Keaven Anderson \email{keaven_anderson@merck.com}
}

}
\keyword{internal}
